safari-600.7-branch/LayoutTests/platform/gtk/editing/input/option-page-up-down-expected.txt

CONSOLE MESSAGE: line 50: cursor should be at line 3, not line 0
Page up/down (option+page up/down on Mac) should move the move cursor and scroll the content in contenteditable elements. This sample covers cursor position move test of a) iframe element containing contenteditable body, and b) content editable div element.
